+++ Game/Move.cs
@@ -1,0 +1,6 @@
+﻿namespace Game
+{
+    public class Move
+    {
+    }
+}
